+++ token.go
@@ -38,7 +38,7 @@ type Tokenizer struct {
 	AllowComments  bool
 	ForceEOF       bool
 	lastChar       uint16
-	Position       int
+	Position       int64
 	lastToken      []byte
 	LastError      error
 	posVarIndex    int
@@ -791,7 +791,7 @@ func (tkn *Tokenizer) scanString(delim uint16, typ int) (int, []byte) {
 			}
 
 			buffer.Write(tkn.buf[start:tkn.bufPos])
-			tkn.Position += (tkn.bufPos - start)
+			tkn.Position += int64(tkn.bufPos - start)
 
 			if tkn.bufPos >= tkn.bufSize {
 				// Reached the end of the buffer without finding a delim or
